backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
openai
psycopg[binary]
sqlalchemy
pydantic
